Fixing access to webjars
Play framework automatically unpacks webjars contents to a fixed folder.
We need to mimic that with Bazel.

A genrule is the way to do that.
We need to unpack the jars in the public "folder" (in the output)
and move the unpacked files to specific output locations that bazel can
recognize.

More info at:
https://www.playframework.com/documentation/2.8.1/AssetsOverview#WebJars
https://docs.bazel.build/versions/main/be/general.html#genrule